Python/Django-Rest/requirements.txt
asgiref==3.2.10
Django==3.0.8
django-jet==1.0.8
djangorestframework==3.11.0
mysql==0.0.2
mysqlclient==2.0.1
pytz==2020.1
sqlparse==0.3.1
